@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 
 import { UiScrollComponent } from './ui-scroll.component';
 import { UiScrollDirective } from './ui-scroll.directive';
 
+const imports = [UiScrollComponent, UiScrollDirective];
+
 @NgModule({
-  declarations: [UiScrollComponent, UiScrollDirective],
-  imports: [CommonModule],
-  exports: [UiScrollDirective]
+  imports: imports,
+  exports: imports
 })
 export class UiScrollModule {}
